extract welcome game flicker component
