+++ modules/ui/src/components/TitlePage/TitlePage.css
@@ -3,7 +3,7 @@
   width: 100%;
 }
 
-.titlePageBoardBoxContainer {
+.titlePageWelcomeContainer {
   position: absolute;
   z-index: 100;
   top: 30px;
